REFACTOR :: Layout 요소 줄임 - layouts 폴더 삭제후 common으로 옮김 layouts안에 불필요하게 많은 layout 요소를 줄임
