+++ apps/user/src/app/form/page.tsx
@@ -1,9 +1,9 @@
-"use client";
+'use client';
 
-import FormLayout from "@/layouts/FormLayout";
+import FormLayout from '@/components/common/Layouts/FormLayout';
 
 const FormPage = () => {
-  return <FormLayout title="지원자 정보">a</FormLayout>;
+    return <FormLayout title="지원자 정보">a</FormLayout>;
 };
 
 export default FormPage;
