+++ apps/user/src/components/common/Layouts/AppLayout.tsx
@@ -1,3 +1,4 @@
+import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import { color } from '@maru/theme';
 import { ReactNode } from 'react';
@@ -6,13 +7,15 @@ import styled from 'styled-components';
 interface PropsType {
     children: ReactNode;
     backgroundColor?: string;
+    padding?: string;
 }
 
-const AppLayout = ({ children, backgroundColor }: PropsType) => {
+const AppLayout = ({ children, backgroundColor = color.white, padding }: PropsType) => {
     return (
         <>
             <Header />
-            <StyledAppLayout style={{ backgroundColor }}>{children}</StyledAppLayout>
+            <StyledAppLayout style={{ backgroundColor, padding }}>{children}</StyledAppLayout>
+            <Footer />
         </>
     );
 };
@@ -21,6 +24,5 @@ export default AppLayout;
 
 const StyledAppLayout = styled.section`
     width: 100vw;
-    min-height: 100%;
-    background-color: ${color.white};
+    height: 100vh;
 `;
