--- apps/user/src/layouts/FaqLayout.tsx
@@ -1,28 +1,0 @@
-import Footer from '@/components/common/Footer';
-import Header from '@/components/common/Header';
-import { color } from '@maru/theme';
-import { ReactNode } from 'react';
-import styled from 'styled-components';
-
-interface PropsType {
-    children: ReactNode;
-}
-
-const FaqLayout = ({ children }: PropsType) => {
-    return (
-        <>
-            <Header />
-            <StyledFaqLayout>{children}</StyledFaqLayout>
-            <Footer />
-        </>
-    );
-};
-
-export default FaqLayout;
-
-const StyledFaqLayout = styled.section`
-    width: 100vw;
-    height: 100vh;
-    background-color: ${color.white};
-    padding: 0px 207px;
-`;
